[531] refactoring smallAmountTooltip
