+++ src/modules/land-works/components/land-works-list-new-summary/index.tsx
@@ -3,7 +3,7 @@ import BigNumber from 'bignumber.js';
 
 import { ReactComponent as AlertIcon } from 'assets/icons/warning.svg';
 import Icon from 'components/custom/icon';
-import SmallAmountTooltip from 'components/custom/smallAmountTooltip';
+import SmallAmountTooltip from 'components/custom/small-amount-tooltip';
 import { Grid } from 'design-system';
 import { getTokenIconName } from 'helpers/helpers';
 import { AssetEntity, PaymentToken } from 'modules/land-works/api';
